deploy/requirements.txt
torch
easyocr
torchvision
supervision>=0.18.0
transformers
ultralytics==8.3.61
azure-identity
numpy
opencv-python
opencv-python-headless
gradio
dill
accelerate
timm
einops==0.8.0
paddlepaddle
paddleocr
ipykernel
Pillow